% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down.R
\name{down}
\alias{down}
\title{Function associated with the OwnFunction distribution}
\usage{
down(x)
}
\arguments{
\item{x}{A numeric value or vector where the function is evaluated}
}
\value{
Returns the value of the function at x
}
\description{
Function associated with the OwnFunction distribution
}
\details{
This function evaluates a piecewise-defined function on the interval (-1, 2).
It is given by:
\deqn{
f(x) = \begin{cases}
\frac{x^2}{3} & -1 < x < 2 \\
0 & \text{otherwise}
\end{cases}
}
}
\examples{
down(1)
down(c(-0.5, 0.5, 1.5))
}
