% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pown.R
\name{pown}
\alias{pown}
\title{Cumulative Distribution Function of the OwnFunction Distribution}
\usage{
pown(x)
}
\arguments{
\item{x}{A numeric value or vector where the CDF is evaluated.}
}
\value{
A numeric vector with the corresponding cumulative probabilities.
}
\description{
Evaluates the cumulative distribution function (CDF) of the OwnFunction
continuous distribution at given values of \code{x}.
}
\details{
The OwnFunction distribution is defined on the interval (-1, 2). Its
cumulative distribution function is:

\deqn{
F(x) =
\begin{cases}
0, & x \leq -1 \\
\frac{x^3 + 1}{9}, & -1 < x < 2 \\
1, & x \geq 2
\end{cases}
}

This function is vectorized and returns values in the interval [0, 1].
}
\examples{
pown(0)
pown(c(-1, 0, 1, 2))

}
