% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rown.R
\name{rown}
\alias{rown}
\title{Random Generator using Inverse Method}
\usage{
rown(n)
}
\arguments{
\item{n}{Integer. Number of random values to generate.}
}
\value{
A numeric vector of length \code{n} with random values following the defined distribution.
}
\description{
Generates random numbers from a continuous distribution using the inverse transform method.
}
\details{
This function uses the inverse of the cumulative distribution function (CDF):
\deqn{F^{-1}(u) = (9u - 1)^{1/3}}
where \eqn{u \sim U(0,1)}.

To avoid numerical problems with negative values, the cube root is computed as:
\deqn{\text{sign}(x)\,|x|^{1/3}}
}
\examples{
# Generate 1000 random values
rown(1000)

}
