% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/down.R
\name{down}
\alias{down}
\title{Probability density function of the OwnFunction distribution}
\usage{
down(x)
}
\arguments{
\item{x}{A numeric value or vector where the PDF is evaluated}
}
\value{
Returns the value of the probability density function at x
}
\description{
Probability density function of the OwnFunction distribution
}
\details{
This function computes the probability density function (PDF) associated
with the custom continuous distribution defined in the package. The function is:
\deqn{
f(x) = \begin{cases}
\frac{x^2}{3} & -1 < x < 2 \\
0 & \text{otherwise}
\end{cases}
}
}
\examples{
down(0)
down(c(-2, -0.5, 0, 0.5, 1.5, 2.5))
}
