% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownChisq.R
\name{ownChisq}
\alias{ownChisq}
\title{Chi-square Goodness-of-Fit Test for a sample of the OwnFunction Distribution
#'}
\usage{
ownChisq(sample)
}
\arguments{
\item{sample}{Numeric vector of observations.}
}
\value{
A named numeric vector with:
\itemize{
  \item statistic : Chi-square statistic
  \item df        : Degrees of freedom
  \item p.val     : p-value
}
}
\description{
Performs a chi-square goodness-of-fit test comparing observed frequencies
from a sample with expected probabilities derived from the distribution:

\deqn{
f(x) =
\begin{cases}
\frac{x^2}{3}, & -1 < x < 2 \\
0, & \text{otherwise}
\end{cases}
}

The interval [-1, 2] is divided into bins of width 0.25.
}
\details{
The CDF is:
\deqn{
F(x) = \frac{x^3}{9} + \frac{1}{9}
}

Expected probabilities are computed as differences of the CDF over each interval.
Observed frequencies are obtained using the \code{fdth} package.
}
\examples{
sample <- runif(100, -1, 2)
ownChisq(sample)

}
