% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pown.R
\name{pown}
\alias{pown}
\title{Cumulative distribution function of the OwnFunction distribution}
\usage{
pown(x)
}
\arguments{
\item{x}{A numeric value or vector where the CDF is evaluated}
}
\value{
Returns the value of the cumulative distribution function at x
}
\description{
Cumulative distribution function of the OwnFunction distribution
}
\details{
This function computes the cumulative distribution function (CDF) associated
with the custom continuous distribution defined in the package. The function is:
\deqn{
F(x) = \begin{cases}
0 & x \leq -1 \\
\frac{x^3 + 1}{9} & -1 < x < 2 \\
1 & x \geq 2
\end{cases}
}
}
\examples{
pown(0)
pown(c(-1,2,5,3))
}
